pyproject.toml
[tool.poetry]
name = "code"
version = "0.1.0"
description = "The compiler backend for Test.it"
authors = ["KhushalJangid <kjangid897@gmail.com>"]
package-mode = false
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.1.2"
django-rest-framework = "^0.1.0"
django-cors-headers = "^4.5.0"
django-channels = "^0.7.0"
djangorestframework-simplejwt = "^5.3.1"
django-unfold = "^0.40.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
